refactor(parquet): inline capped Auto tests
